Fix: Firebase Sentry
