+++ src/pages/_app.tsx
@@ -1,4 +1,3 @@
-import * as Sentry from '@sentry/browser';
 import { ApolloClient, ApolloProvider, InMemoryCache } from '@apollo/client';
 import {
     AppContainer,
@@ -12,7 +11,8 @@ import { PrismicDataProvider } from '../libs/Prismic/components/PrismicDataProvi
 import { Provider, useSelector } from 'react-redux';
 import { WagmiConfig, useAccount } from 'wagmi';
 import { getCookie, hasCookie } from 'cookies-next';
-import { getMessaging, getToken, isSupported } from 'firebase/messaging';
+import { handleFirebaseServiceWorker } from 'src/hooks/firebase/handleFirebaseServiceWorker';
+import { isSupported } from 'firebase/messaging';
 import { registerFirebaseSW } from 'src/hooks/useServiceWorker';
 import {
     selectCurrentUser,
@@ -30,7 +30,6 @@ import SEO from '../components/SEO';
 import Sidebar from '../components/Sidebar';
 import WrapperProvider from '../components/WrapperProvider';
 import config from '../../config';
-import firebaseApp from 'src/utils/firebase/firebase';
 import modals from '../modals';
 import useGuard from '../hooks/useGuard';
 import type { AppProps } from 'next/app';
@@ -60,52 +59,12 @@ const InnerApp = (props: AppProps) => {
         }
 
         if (isConnected && (signature || eip712_signature)) {
-            const handleFirebaseServiceWorker = async () => {
-                if (isSupported) {
-                    const messaging = getMessaging(firebaseApp);
-                    const permission = await Notification.requestPermission();
-
-                    if (permission === 'granted') {
-                        const currentToken = await getToken(messaging, {
-                            vapidKey: config.fbVapidKey
-                        });
-
-                        if (currentToken) {
-                            // Send Firebase token to endpoint /user
-                            try {
-                                await fetch(`${config.baseApiUrl}/users`, {
-                                    body: JSON.stringify({
-                                        appPNT: currentToken
-                                    }),
-                                    headers: {
-                                        Accept: 'application/json',
-                                        Authorization: `Bearer ${getCookie(
-                                            'AUTH_TOKEN'
-                                        ).toString()}`,
-                                        'Content-Type': 'application/json',
-                                        eip712signature: eip712_signature,
-                                        eip712value: eip712_message,
-                                        message,
-                                        signature
-                                    },
-                                    method: 'PUT'
-                                });
-                            } catch (e: any) {
-                                console.log(e);
-                                Sentry.captureException(e);
-                            }
-                        } else {
-                            console.log(
-                                'No registration token available. Request permission to generate one.'
-                            );
-                        }
-                    } else {
-                        console.log('Notification permission not granted');
-                    }
-                }
-            };
-
-            handleFirebaseServiceWorker();
+            handleFirebaseServiceWorker(
+                eip712_signature,
+                eip712_message,
+                message,
+                signature
+            );
         }
     }, [isConnected, signature, eip712_signature]);
 
@@ -172,7 +131,15 @@ const App = (props: AppProps) => {
     }
 
     useEffect(() => {
-        registerFirebaseSW();
+        const registerFirebaseServiceWorker = async () => {
+            const hasFirebaseSupport = await isSupported();
+
+            if (hasFirebaseSupport) {
+                registerFirebaseSW();
+            }
+        };
+
+        registerFirebaseServiceWorker();
     }, []);
 
     return (
